Rename package and add package path
