+++ 안병규/week1/소형기관차/Main.java
