+++ 안병규/week2/가장긴증가하는부분수열2/Main.java
@@ -1,4 +1,4 @@
-package 가장긴증가하는부분수열2;
+package week2.가장긴증가하는부분수열2;
 
 import java.io.BufferedReader;
 import java.io.IOException;
