+++ 안병규/week3/BOJ_다리만들기2/Main.java
@@ -1,3 +1,5 @@
+package week3.BOJ_다리만들기2;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
